+++ AddOns/GSE/API/translator.lua
@@ -7,7 +7,8 @@ local L = GSE.L
 
 --- GSE.TranslateSequence will translate from local spell name to spell id and back again.\
 -- Mode of "STRING" will return local names where mode "ID" will return id's
-function GSE.TranslateSequence(sequence, sequenceName, mode)
+-- dropAbsolute will remove "$$" from the start of lines.
+function GSE.TranslateSequence(sequence, sequenceName, mode, dropAbsolute)
     GSE.PrintDebugMessage("GSE.TranslateSequence  Mode: " .. mode, GNOME)
 
     for k, v in ipairs(sequence) do
@@ -17,7 +18,7 @@ function GSE.TranslateSequence(sequence, sequenceName, mode)
 
     if not GSE.isEmpty(sequence.KeyRelease) then
         for k, v in pairs(sequence.KeyRelease) do
-            sequence.KeyRelease[k] = GSE.TranslateString(v, mode)
+            sequence.KeyRelease[k] = GSE.TranslateString(v, mode, dropAbsolute)
         end
     else
         GSE.PrintDebugMessage("empty Keyrelease in translate", Statics.Translate)
@@ -26,7 +27,7 @@ function GSE.TranslateSequence(sequence, sequenceName, mode)
         GSE.PrintDebugMessage("Keypress has stuff in translate", Statics.Translate)
         for k, v in pairs(sequence.KeyPress) do
             -- Translate KeyRelease
-            sequence.KeyPress[k] = GSE.TranslateString(v, mode)
+            sequence.KeyPress[k] = GSE.TranslateString(v, mode, dropAbsolute)
         end
     else
         GSE.PrintDebugMessage("empty Keypress in translate", Statics.Translate)
@@ -35,7 +36,7 @@ function GSE.TranslateSequence(sequence, sequenceName, mode)
         GSE.PrintDebugMessage("Keypress has stuff in translate", Statics.Translate)
         for k, v in pairs(sequence.PreMacro) do
             -- Translate KeyRelease
-            sequence.PreMacro[k] = GSE.TranslateString(v, mode)
+            sequence.PreMacro[k] = GSE.TranslateString(v, mode, dropAbsolute)
         end
     else
         GSE.PrintDebugMessage("empty Keypress in translate", Statics.Translate)
@@ -44,7 +45,7 @@ function GSE.TranslateSequence(sequence, sequenceName, mode)
         GSE.PrintDebugMessage("Keypress has stuff in translate", Statics.Translate)
         for k, v in pairs(sequence.PostMacro) do
             -- Translate KeyRelease
-            sequence.PostMacro[k] = GSE.TranslateString(v, mode)
+            sequence.PostMacro[k] = GSE.TranslateString(v, mode, dropAbsolute)
         end
     else
         GSE.PrintDebugMessage("empty Keypress in translate", Statics.Translate)
@@ -102,11 +103,16 @@ function GSE.ProcessVariables(lines, variableTable)
     return returnLines
 end
 
-function GSE.TranslateString(instring, mode, cleanNewLines)
+function GSE.TranslateString(instring, mode, cleanNewLines, dropAbsolute)
     instring = GSE.UnEscapeString(instring)
     GSE.PrintDebugMessage("Entering GSE.TranslateString with : \n" .. instring .. "\n " .. mode, GNOME)
     local output = ""
     if not GSE.isEmpty(instring) then
+        local absolute = false
+        if instring:find('$$', 1, true) then
+            GSE.PrintDebugMessage("Setting Absolute", GNOME)
+            absolute = true
+        end
         if GSE.isEmpty(string.find(instring, '--', 1, true)) then
             for cmd, etc in string.gmatch(instring or '', '/(%w+)%s+([^\n]+)') do
                 GSE.PrintDebugMessage("cmd : \n" .. cmd .. " etc: " .. etc, GNOME)
@@ -130,7 +136,7 @@ function GSE.TranslateString(instring, mode, cleanNewLines)
                         output = output .. "!"
                     end
                     local foundspell, returnval = GSE.TranslateSpell(etc, mode,
-                                                      (cleanNewLines and cleanNewLines or false))
+                                                      (cleanNewLines and cleanNewLines or false), absolute)
                     if foundspell then
                         output = output .. GSEOptions.KEYWORD .. returnval .. Statics.StringReset
                     else
@@ -156,7 +162,7 @@ function GSE.TranslateString(instring, mode, cleanNewLines)
                                 output = output .. "!"
                             end
                             local foundspell, returnval = GSE.TranslateSpell(uetc, mode,
-                                                              (cleanNewLines and cleanNewLines or false))
+                                                              (cleanNewLines and cleanNewLines or false), asbolute)
                             output = output .. GSEOptions.KEYWORD .. returnval .. Statics.StringReset .. ", "
                         end
                         output = output .. ";"
@@ -192,6 +198,13 @@ function GSE.TranslateString(instring, mode, cleanNewLines)
                 output = string.gsub(output, "/" .. v, GSEOptions.WOWSHORTCUTS .. "/" .. v .. Statics.StringReset)
             end
         end
+
+        if GSE.isEmpty(dropAbsolute) then
+            dropAbsolute = false
+        end
+        if  absolute and not dropAbsolute then
+            output = "$$" .. output
+        end
     elseif cleanNewLines then
         output = output .. instring
     end
@@ -206,7 +219,7 @@ function GSE.TranslateString(instring, mode, cleanNewLines)
     return output
 end
 
-function GSE.TranslateSpell(str, mode, cleanNewLines)
+function GSE.TranslateSpell(str, mode, cleanNewLines, absolute)
     local output = ""
     local found = false
     -- Check for cases like /cast [talent:7/1] Bladestorm;[talent:7/3] Dragon Roar
@@ -238,7 +251,7 @@ function GSE.TranslateSpell(str, mode, cleanNewLines)
             etc = string.match(etc, "^%s*(.-)%s*$")
         end
 
-        local foundspell = GSE.GetSpellId(etc, mode)
+        local foundspell = GSE.GetSpellId(etc, mode, absolute)
         -- print("Foudn Spell: " .. foundspell .. " etc:" .. etc .. " mode:" .. mode .. " str:" .. str)
         if foundspell then
             GSE.PrintDebugMessage("Translating Spell ID : " .. etc .. " to " .. foundspell, GNOME)
@@ -692,7 +705,7 @@ function GSE.ClassicGetSpellInfo(spellID)
 end
 
 --- Converts a string spell name to an id and back again.
-function GSE.GetSpellId(spellstring, mode)
+function GSE.GetSpellId(spellstring, mode, absolute)
     if GSE.isEmpty(mode) then
         mode = ""
     end
@@ -709,8 +722,14 @@ function GSE.GetSpellId(spellstring, mode)
         if GSE.GameMode ~= 1 then
             -- If we are not in classic
             -- Check for overrides like Crusade and Avenging Wrath.
-            if not GSE.isEmpty(Statics.BaseSpellTable[returnval]) then
-                returnval = Statics.BaseSpellTable[returnval]
+            if not absolute then
+                if FindBaseSpellByID(returnval) then
+                    returnval = FindBaseSpellByID(returnval)
+                end
+                -- Still need Heart of Azeroth overrides.
+                if not GSE.isEmpty(Statics.BaseSpellTable[returnval]) then
+                    returnval = Statics.BaseSpellTable[returnval]
+                end
             end
         end
     end