+++ AddOns/GSE/Localization/ModL_zhCN.lua
@@ -611,6 +611,8 @@ L["You need to reload the User Interface to complete this task.  Would you like 
 --[[Translation missing --]]
 L["Your ClassID is "] = "Your ClassID is "
 L["Your current Specialisation is "] = "你当前的专精是"
+--[[Translation missing --]]
+L["Your sequence name was longer than 27 characters.  It has been shortened from %s to %s so that your macro will work."] = "Your sequence name was longer than 27 characters.  It has been shortened from %s to %s so that your macro will work."