+++ AddOns/Altoholic/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Altoholic Classic - Fork by Teelo
 
-## [Classic-v1.13.5-050](https://github.com/teelolws/Altoholic-Classic/tree/Classic-v1.13.5-050) (2020-10-17)
-[Full Changelog](https://github.com/teelolws/Altoholic-Classic/compare/Classic-v1.13.5-049...Classic-v1.13.5-050) [Previous Releases](https://github.com/teelolws/Altoholic-Classic/releases)
+## [Classic-v1.13.5-052](https://github.com/teelolws/Altoholic-Classic/tree/Classic-v1.13.5-052) (2020-10-18)
+[Full Changelog](https://github.com/teelolws/Altoholic-Classic/compare/Classic-v1.13.5-051...Classic-v1.13.5-052) [Previous Releases](https://github.com/teelolws/Altoholic-Classic/releases)
 
-- Version ID fix  
+- Update DataStore\_Containers.lua